+++ tests/Middleware/RouteScopesTest.php
@@ -3,7 +3,7 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the GNU Affero General Public License v3.0.
 // See the LICENCE file in the repository root for full licence text.
 
-namespace Tests;
+namespace Tests\Middleware;
 
 use App\Http\Middleware\RequireScopes;
 use App\Libraries\RouteScopesHelper;
@@ -12,6 +12,7 @@ use App\Models\Changelog;
 use App\Models\Comment;
 use App\Models\UpdateStream;
 use Route;
+use Tests\TestCase;
 
 class RouteScopesTest extends TestCase
 {
